Add HUBOT_BRAIN_DIR instructions to README
This commit adds previously missing instructions for setting the
`HUBOT_BRAIN_DIR` environment variable when creating the docker
container. Without this setting the brain file will not be saved in the
expected location for binding it back to a local volume. Failure to set
this variable will result in the loss of the `hubot_brain.json` file and
all persistent data will be lost.