ASC-1528 Add create update submodule pr script
This commit adds a script for creating PRs to update submodules based on
the given repo and branch.  The script responds with the PR url upon
success or will respond with any error encountered during the process.

This script requires that a private ssh key be added to the docker build
as a `--build-arg`.